Use conditional imports instead of conditional exports for FileIO.
