+++ packages/devtools_app/lib/src/config_specific/file/_fake_file.dart
@@ -1,28 +1,9 @@
-// Copyright 2019 The Chromium Authors. All rights reserved.
+// Copyright 2020 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'file_io.dart';
+import 'file.dart';
 
-/// Abstracted file system access for Flutter Web.
-class FileSystem implements FileIO {
-  // TODO(terry): Implement Web based file IO.
-
-  /// Key is filename and value is content of the file.
-  final Map<String, String> _files = {};
-
-  @override
-  void writeStringToFile(String filename, String contents) {
-    _files.putIfAbsent(filename, () => contents);
-  }
-
-  @override
-  String readStringFromFile(String filename) =>
-      _files.containsKey(filename) ? _files[filename] : null;
-
-  @override
-  List<String> list({String prefix}) => _files.keys.toList();
-
-  @override
-  bool deleteFile(String filename) => _files.remove(filename) != null;
+FileIO createFileSystem() {
+  throw Exception('Attempting to create FileSystem for unrecognized platform.');
 }
